load js source code with //go:embed
