+++ js/attribute.js
@@ -1,0 +1,3 @@
+function attribute(n) {
+    return this[n];
+}
