+++ js/blur.js
@@ -1,0 +1,4 @@
+function blur() {
+    this.blur();
+    return true;
+}
