+++ js/text.js
@@ -1,0 +1,6 @@
+function text() {
+    if (this.offsetWidth || this.offsetHeight || this.getClientRects().length) {
+        return this.innerText;
+    }
+    return '';
+}
