+++ js/textContent.js
@@ -1,0 +1,3 @@
+function textContent() {
+    return this.textContent;
+}
