+++ js/visible.js
@@ -1,0 +1,3 @@
+function visible() {
+    return Boolean(this.offsetWidth || this.offsetHeight || this.getClientRects().length);
+}
